docTransformer/pyproject.toml
[tool.poetry]
name = "ts-expert2"
version = "0.1.0"
description = ""
authors = ["Madung2 <tuliphan91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
pandas = "^2.2.2"
python-docx = "^1.1.2"
fuzzywuzzy = "^0.18.0"
konlpy = "^0.6.0"
camelot-py = "^0.11.0"
opencv-python = "^4.10.0.84"
python-mecab-ko = "^1.3.7"
python-levenshtein = "^0.25.1"
django-cors-headers = "^4.4.0"
spacy = "^3.7.5"
pdf2docx = "^0.5.8"
pypdf2 = "^3.0.1"
celery = "^5.4.0"
redis = "^5.0.8"
pypandoc = "^1.13"
xlsxwriter = "^3.2.0"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
